feat : make minishell utils file #5
